feat: hide uninstall btn for dev and pluginmarket

+++ src/plugin-manager/index.tsx
@@ -85,7 +85,7 @@ export async function initPluginManager() {
 	});
 }
 
-export let onPluginLoaded = (_: typeof loadedPlugins) => {};
+export let onPluginLoaded = (_: typeof loadedPlugins) => { };
 
 const PluginManager: React.FC = () => {
 	const [selectedPlugin, setSelectedPlugin] = React.useState<NCMPlugin | null>(
@@ -190,38 +190,46 @@ const PluginManager: React.FC = () => {
 													{loadPlugin.manifest.name}
 												</span>
 												{/* rome-ignore lint/a11y/useKeyWithClickEvents: <explanation> */}
-												<span
-													className="plugin-uninstall-btn"
-													onClick={async (e) => {
-														e.stopPropagation();
-														const pluginFilePath =
-															await BetterNCM.fs.readFileText(
-																`${loadPlugin.pluginPath}/.plugin.path.meta`,
-															);
-														await BetterNCM.fs.remove(pluginFilePath);
+												{
+													(!loadPlugin.pluginPath.includes("./plugins_dev") && loadPlugin.manifest.name !== "PluginMarket") &&
+													(
+														<span
+															className="plugin-uninstall-btn"
+															onClick={async (e) => {
+																e.stopPropagation();
+																const pluginFilePath =
+																	await BetterNCM.fs.readFileText(
+																		`${loadPlugin.pluginPath}/.plugin.path.meta`,
+																	);
+																if (pluginFilePath.length > 1) {
+																	await BetterNCM.fs.remove(pluginFilePath);
+																	await BetterNCM.app.reloadPlugins();
+																	BetterNCM.reload();
+																}
+															}}
+														>
+															<svg
+																xmlns="http://www.w3.org/2000/svg"
+																width={24}
+																height={24}
+																viewBox="0 0 24 24"
+																fill="none"
+																stroke="currentColor"
+																strokeWidth={2}
+																strokeLinecap="round"
+																strokeLinejoin="round"
+																className="feather feather-trash-2"
+															>
+																<polyline points="3 6 5 6 21 6" />
+																<path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" />
+																<line x1={10} y1={11} x2={10} y2={17} />
+																<line x1={14} y1={11} x2={14} y2={17} />
+															</svg>
+														</span>
+													)
+												}
 
-														await BetterNCM.app.reloadPlugins();
-														BetterNCM.reload();
-													}}
-												>
-													<svg
-														xmlns="http://www.w3.org/2000/svg"
-														width={24}
-														height={24}
-														viewBox="0 0 24 24"
-														fill="none"
-														stroke="currentColor"
-														strokeWidth={2}
-														strokeLinecap="round"
-														strokeLinejoin="round"
-														className="feather feather-trash-2"
-													>
-														<polyline points="3 6 5 6 21 6" />
-														<path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" />
-														<line x1={10} y1={11} x2={10} y2={17} />
-														<line x1={14} y1={11} x2={14} y2={17} />
-													</svg>
-												</span>
+
 											</div>
 										);
 									})}
